+++ SCollector/SCollector/HudObject.cpp
@@ -8,22 +8,22 @@ HudObject::HudObject(const std::string& image, const sf::Vector2f& pos)
 	value = 100;
 	texture.loadFromFile(image);
 	sprite.setTexture(texture);
-	sprite.setOrigin(0,sprite.getTextureRect().height/2);
+	sprite.setOrigin(0,sprite.getTextureRect().height/2.0f);
 	sprite.setPosition(pos.x + sprite.getOrigin().x, pos.y - sprite.getOrigin().y);
 	rect.setFillColor(sf::Color::Red);
 	
-	rect.setSize(sf::Vector2f(sprite.getTextureRect().width,sprite.getTextureRect().height));
+	rect.setSize(sf::Vector2f((float)sprite.getTextureRect().width,(float)sprite.getTextureRect().height));
 	rect.setPosition(pos.x,pos.y-rect.getSize().y);
 }
 
 float HudObject::getWidth()
 {
-	return sprite.getTextureRect().width;
+	return (float)sprite.getTextureRect().width;
 }
 
 float HudObject::getHeight()
 {
-	return sprite.getTextureRect().height;
+	return (float)sprite.getTextureRect().height;
 }
 
 void HudObject::changeFill(int final)
@@ -31,8 +31,8 @@ void HudObject::changeFill(int final)
 	int delta = final - value;
 	value = final;
 
-	rect.setSize(sf::Vector2f(rect.getSize().x,sprite.getTextureRect().height*value/100.0));
-	rect.setPosition(rect.getPosition().x,rect.getPosition().y-sprite.getTextureRect().height*delta/100.0);
+	rect.setSize(sf::Vector2f(rect.getSize().x,sprite.getTextureRect().height*value/100.0f));
+	rect.setPosition(rect.getPosition().x,rect.getPosition().y-sprite.getTextureRect().height*delta/100.0f);
 }
 
 void HudObject::draw(sf::RenderTarget& target, sf::RenderStates states) const
